requirements.txt
arabic_reshaper==2.0.15
python-bidi==0.4.2
setuptools>=65.0.0